refactor(Network): 修改消息头长度属性类型为ushort
将PackageHeaderLength属性从uint改为ushort以优化内存使用，因为消息头长度不会超过ushort的最大值

+++ GameFrameX.NetWork.Message/ClientMessageDecoderHandler.cs
@@ -16,8 +16,7 @@ public sealed class ClientMessageDecoderHandler : DefaultMessageDecoderHandler
     /// <summary>
     /// 消息头长度
     /// </summary>
-    public override uint PackageHeaderLength { get; } = sizeof(uint) + sizeof(byte) + sizeof(byte) + sizeof(int) + sizeof(int);
-
+    public override ushort PackageHeaderLength { get; } = sizeof(uint) + sizeof(byte) + sizeof(byte) + sizeof(int) + sizeof(int);
 
     /// <summary>
     /// 消息解码
